fixed some kinks

+++ screenbolus.h
@@ -23,9 +23,9 @@ private:
 
 signals:
     void sendToHome();
-    void sendConfirmBolus(int, int, int, double, double);
+    void sendConfirmBolus(int, int, double, double, double);
     void sendCalcUnits(double, double);
-    void sendCalcExtended(int, int, int, double, double);
+    void sendCalcExtended(int, int, double, double, double);
 
 private slots:
     void goToHome();
